+++ bundler/macos/bundle.sh
@@ -220,7 +220,7 @@ $SHOULD_MAKE_DMG \
 	-format UDZO \
 	-imagekey zlib-level=9 \
 	-srcfolder "$OUTPUT_CONTAINER_DIR" \
-	-o "$BIN_MACOS_DIR/${APP_NAME}_${EFFECTIVE_VERSION}.dmg"
+	-o "$BIN_MACOS_DIR/${APP_NAME}.dmg"
 
 $SHOULD_OPEN && open "$OUTPUT_DIR"